# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TrendRadar",
    "description": "AI \u9a71\u52a8\u7684\u5168\u7403\u79d1\u6280\u70ed\u70b9\u53d1\u73b0\u5e73\u53f0",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/collections/folders": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "\u6211\u7684\u6536\u85cf\u5939\u5217\u8868",
        "operationId": "list_folders_api_v1_collections_folders_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/FolderResponse"
                  },
                  "type": "array",
                  "title": "Response List Folders Api V1 Collections Folders Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "\u65b0\u5efa\u6536\u85cf\u5939\uff08\u226450\uff09",
        "operationId": "create_folder_api_v1_collections_folders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FolderCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/collections/items": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "\u6211\u7684\u6536\u85cf\u6761\u76ee",
        "operationId": "list_items_api_v1_collections_items_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "folder_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Folder Id"
            }
          },
          {
            "name": "read_status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReadStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Read Status"
            }
          },
          {
            "name": "user_tag",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 20
                },
                {
                  "type": "null"
                }
              ],
              "title": "User Tag"
            }
          },
          {
            "name": "keyword",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 200
                },
                {
                  "type": "null"
                }
              ],
              "title": "Keyword"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "-createdAt / -updatedAt / -readAt",
              "default": "-createdAt",
              "title": "Sort"
            },
            "description": "-createdAt / -updatedAt / -readAt"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "\u9875\u7801\uff0c\u4ece 1 \u5f00\u59cb",
              "default": 1,
              "title": "Page"
            },
            "description": "\u9875\u7801\uff0c\u4ece 1 \u5f00\u59cb"
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "\u6bcf\u9875\u6761\u6570\uff0c\u6700\u5927 100",
              "default": 20,
              "title": "Size"
            },
            "description": "\u6bcf\u9875\u6761\u6570\uff0c\u6700\u5927 100"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ItemResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "collection"
        ],
        "summary": "\u6536\u85cf\u4e00\u4e2a\u4e8b\u4ef6\uff08\u226410000\uff09",
        "operationId": "create_item_api_v1_collections_items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ItemCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ItemResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/collections/stats": {
      "get": {
        "tags": [
          "collection"
        ],
        "summary": "\u6211\u7684\u6536\u85cf\u7edf\u8ba1",
        "operationId": "get_stats_api_v1_collections_stats_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatsResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CategoryCount": {
        "properties": {
          "category": {
            "type": "string",
            "title": "Category"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "category",
          "count"
        ],
        "title": "CategoryCount"
      },
...
      "EventBrief": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summaryOneLine": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summaryoneline"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "recommendIndex": {
            "type": "number",
            "title": "Recommendindex"
          },
          "sourceCount": {
            "type": "integer",
            "title": "Sourcecount"
          },
          "lastSeenAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastseenat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "recommendIndex",
          "sourceCount"
        ],
        "title": "EventBrief",
        "description": "\u6761\u76ee\u5217\u8868\u91cc\u5d4c\u5165\u7684\u4e8b\u4ef6\u7b80\u8981\uff08\u907f\u514d\u524d\u7aef\u4e8c\u6b21\u8bf7\u6c42\uff09\u3002"
      },
...
      "FolderCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "FolderCreateRequest"
      },
      "FolderResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "sortOrder": {
            "type": "integer",
            "title": "Sortorder"
          },
          "isDefault": {
            "type": "boolean",
            "title": "Isdefault"
          },
          "itemCount": {
            "type": "integer",
            "title": "Itemcount"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "sortOrder",
          "isDefault",
          "itemCount",
          "createdAt",
          "updatedAt"
        ],
        "title": "FolderResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ItemCreateRequest": {
        "properties": {
          "eventId": {
            "type": "integer",
            "title": "Eventid"
          },
          "folderId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Folderid"
          },
          "note": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 20000
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "userTags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Usertags"
          },
          "readStatus": {
            "$ref": "#/components/schemas/ReadStatus",
            "default": "UNREAD"
          }
        },
        "type": "object",
        "required": [
          "eventId"
        ],
        "title": "ItemCreateRequest"
      },
      "ItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "folderId": {
            "type": "integer",
            "title": "Folderid"
          },
          "folderName": {
            "type": "string",
            "title": "Foldername"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "userTags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Usertags"
          },
          "readStatus": {
            "$ref": "#/components/schemas/ReadStatus"
          },
          "readAt": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Readat"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "title": "Createdat"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "title": "Updatedat"
          },
          "event": {
            "$ref": "#/components/schemas/EventBrief"
          }
        },
        "type": "object",
        "required": [
          "id",
          "folderId",
          "folderName",
          "readStatus",
          "createdAt",
          "updatedAt",
          "event"
        ],
        "title": "ItemResponse"
      },
...
      "MonthCount": {
        "properties": {
          "month": {
            "type": "string",
            "title": "Month"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "month",
          "count"
        ],
        "title": "MonthCount"
      },
...
      "Page_ItemResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ItemResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "pages": {
            "type": "integer",
            "title": "Pages"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "size",
          "pages"
        ],
        "title": "Page[ItemResponse]"
      },
...
      "ReadStatus": {
        "type": "string",
        "enum": [
          "UNREAD",
          "LATER",
          "READ"
        ],
        "title": "ReadStatus",
        "description": "\u6761\u76ee\u9605\u8bfb\u72b6\u6001\u3002"
      },
...
      "StatsResponse": {
        "properties": {
          "totalItems": {
            "type": "integer",
            "title": "Totalitems"
          },
          "unreadCount": {
            "type": "integer",
            "title": "Unreadcount"
          },
          "laterCount": {
            "type": "integer",
            "title": "Latercount"
          },
          "readCount": {
            "type": "integer",
            "title": "Readcount"
          },
          "folderCount": {
            "type": "integer",
            "title": "Foldercount"
          },
          "byCategory": {
            "items": {
              "$ref": "#/components/schemas/CategoryCount"
            },
            "type": "array",
            "title": "Bycategory"
          },
          "recentMonths": {
            "items": {
              "$ref": "#/components/schemas/MonthCount"
            },
            "type": "array",
            "title": "Recentmonths"
          }
        },
        "type": "object",
        "required": [
          "totalItems",
          "unreadCount",
          "laterCount",
          "readCount",
          "folderCount"
        ],
        "title": "StatsResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}